requirements.txt
# Supabase SDK
supabase>=2.13.0

# Async support
asyncio-compat>=0.1.1
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Image processing
Pillow>=10.2.0
python-magic>=0.4.27

# Progress tracking
tqdm>=4.66.0

# Utilities
pydantic>=2.6.0
python-slugify>=8.0.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Code quality
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# Logging
loguru>=0.7.2
